backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
alembic>=1.11.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.2.0
pytest-asyncio>=0.22.0
google-generativeai>=0.3.0
